Combine nav and content reducers (i was stupid earlier)
