+++ client/src/store/reducer/index.js
@@ -1,11 +1,9 @@
 import { combineReducers } from 'redux';
 
-import nav from './nav'
 import content from './content'
 import auth from './auth'
 
 export default combineReducers({
     content,
-    nav,
     auth
 });
